Fixed handling of exceptions due to CB downtime
Carbon Black's API actually catches ConnectionError and then raises it as an ApiError - added catching of the ApiError containing 'network connection error' message.

Also removed .format() use on StatusMessage in "Catch all other exceptions and abort" region. This would have failed and thrown an AttributeError exception, as format() is not attribute of StatusMessage().